renamed "mother" related classes and packages to "factory" to use a more common naming convention
