+++ src/test/java/com/daugherty/movie/collector/test/controller/MoviesControllerTests.java
@@ -19,7 +19,7 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import java.util.List;
 
-import static com.daugherty.movie.collector.test.mother.MovieMother.*;
+import static com.daugherty.movie.collector.test.factory.MovieFactory.*;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
